IAv3

+++ UI.h
@@ -34,6 +34,6 @@ void Clear();
 Vec2 World2Screen(Vec2 pos, Vec2 playerPos);
 
 void drawDebugLine(Vec2 start, Vec2 end, char r, char g, char b);
-void drawDebugCircle(short x, short y, short radius, char r, char g, char b);
+void drawDebugCircle(int x, int y, short radius, char r, char g, char b);
 
 #endif
